+++ private/chain_queries_fields/index.js
@@ -1,0 +1,21 @@
+'use strict'
+
+const abi = require('./abi')
+const account = require('./account')
+const accounts_by_authorizers = require('./accounts_by_authorizers')
+const block = require('./block')
+const currency_balance = require('./currency_balance')
+const currency_stats = require('./currency_stats')
+const info = require('./info')
+const producers = require('./producers')
+
+module.exports = {
+  abi,
+  account,
+  accounts_by_authorizers,
+  block,
+  currency_balance,
+  currency_stats,
+  info,
+  producers
+}