+++ private/eosio_types/boolean_type.js
@@ -2,9 +2,10 @@
 const { GraphQLScalarType } = require('graphql')
 
 const boolean_type = new GraphQLScalarType({
-  description: `\`Boolean type\``,
+  description: '`Boolean type` true=1 or false=0',
   name: 'bool',
-  parseValue: boolean => boolean
+  parseValue: boolean => boolean,
+  serialize: boolean => (boolean ? true : false)
 })
 
 module.exports = boolean_type