+++ changelog.md
@@ -11,8 +11,10 @@
 
 ## Minor
 
+- Added chain queries and mutations.
 - Added no query argument option, where the smart contract has no query.
 - Can now handle account names that begin with numerical values.
+- Optional arguments supported.
 
 ## Patch